Cleanup typing

+++ src/skrough/typing.py
@@ -2,6 +2,6 @@ from typing import Callable, Optional, Union
 
 import numpy as np
 
-ChaosMeasure = Callable[[np.ndarray, np.int_], float]
+ChaosMeasure = Callable[[np.ndarray, int], float]
 
 RandomState = Optional[Union[int, np.random.SeedSequence, np.random.Generator]]
